# backend/api/swagger.yaml
openapi: 3.0.3
info:
  title: Fantasm API
  version: 2.0.0
  description: |
    JSON API of the Fantasm backend. Every route lives under `/api`.

    **Authentication.** Sign-in is OIDC only (Google, Microsoft Entra). A successful callback sets an
    opaque, HttpOnly, SameSite=Lax session cookie (`__Host-fantasm_session` over HTTPS). There are no passwords.

    **CSRF.** Unsafe methods (POST, PUT, PATCH, DELETE) must come from the site's own origin: a request with a
    foreign `Origin` or `Sec-Fetch-Site: cross-site` is answered with 403.

    **Errors.** Every failure has the same body, see `Error`, including 404 for an unknown path and 405 (with `Allow`) for a wrong method. Hidden, deleted and nonexistent content all
    answer 404 identically, so existence is never revealed. Unexpected failures are a generic 500 with a `requestId`.

    **Pagination.** Lists are keyset-paginated: pass the previous page's `nextCursor` as `cursor`. Cursors are
    signed and bound to the listing they came from. `limit` is clamped to the documented maximum.

    **Limits.** Request bodies are capped (1 MiB), unknown JSON fields are rejected (400), and every route is
    rate limited per user (signed in) or per address (anonymous). Limited answers carry `Retry-After`.
servers:
  - url: /
security: []
...
paths:
...
  /api/ideas/{slug}/thread:
    get:
      tags: [discussion]
      summary: Discussion thread of an idea
      description: A tree of at most 5000 posts, depth capped at 8 (deeper replies attach to the parent's parent). Deleted posts remain as empty tombstones so replies keep their parent.
      parameters:
        - $ref: '#/components/parameters/Slug'
      responses:
        '200':
          description: The thread.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Thread' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/ideas/{slug}/posts:
    post:
      tags: [discussion]
      summary: Write a post or reply
      description: Markdown, sanitized on write (no raw HTML, no images, at most 10 links, 5000 characters). Ten per minute per user.
      security: [{ session: [] }]
      parameters:
        - $ref: '#/components/parameters/Slug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [body]
              properties:
                body: { type: string, maxLength: 5000 }
                parentId: { type: string, format: uuid }
      responses:
        '201':
          description: The new post.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Post' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Validation' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/posts/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    patch:
      tags: [discussion]
      summary: Edit own post (within 15 minutes)
      security: [{ session: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [body]
              properties:
                body: { type: string, maxLength: 5000 }
      responses:
        '200':
          description: The edited post.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Post' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Validation' }
        '429': { $ref: '#/components/responses/RateLimited' }
    delete:
      tags: [discussion]
      summary: Delete a post (author or staff)
      description: Idempotent soft delete.
      security: [{ session: [] }]
      responses:
        '204': { description: "Deleted." }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: __Host-fantasm_session
      description: Opaque session token set by the OIDC callback. Only its SHA-256 hash is stored.
  parameters:
...
    Slug:
      name: slug
      in: path
      required: true
      schema: { type: string, pattern: '^[a-z0-9]+(-[a-z0-9]+)*$', maxLength: 90 }
...
  responses:
    Unauthorized:
      description: Not signed in, or the session expired.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: Signed in but not allowed, or a cross-origin request.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: No such resource, or one the viewer may not know about.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    Validation:
      description: The input is well-formed but not acceptable; `fields` names what to fix.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    RateLimited:
      description: Too many requests.
      headers:
        Retry-After: { schema: { type: integer }, description: Seconds to wait. }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Error:
      type: object
      required: [error, code]
      properties:
        error: { type: string }
        code:
          type: string
          enum: [invalid_request, unauthorized, forbidden, not_found, conflict, rate_limited, payload_too_large, validation_failed, unsupported_media_type, method_not_allowed, not_implemented, internal_error]
        fields: { type: object, additionalProperties: { type: string } }
        requestId: { type: string }
...
    PostAuthor:
      type: object
      properties:
        handle: { type: string }
        name: { type: string }
        verified: { type: boolean }
    Post:
      type: object
      properties:
        id: { type: string, format: uuid }
        author: { $ref: '#/components/schemas/PostAuthor' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        html: { type: string, description: "Sanitized HTML; empty when deleted." }
        text: { type: string }
        deleted: { type: boolean }
        replies: { type: array, items: { $ref: '#/components/schemas/Post' } }
    Thread:
      type: object
      properties:
        id: { type: string, format: uuid }
        ideaSlug: { type: string }
        count: { type: integer, description: "Posts that are not deleted." }
        posts: { type: array, items: { $ref: '#/components/schemas/Post' } }
...